LayoutTests/fast/mediastream/RTCPeerConnection-state-expected.txt

Tests the RTCPeerConnection states.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc = new webkitRTCPeerConnection({iceServers:[]}, null); did not throw exception.
PASS pc.readyState is "new"
PASS stateChanged was called
PASS pc.readyState is "closed"
PASS successfullyParsed is true

TEST COMPLETE
